refactor: split repo contracts into Reader + Writer roles (flysystem-style)
Borrows the role-interface split from league/flysystem: each repo
contract becomes the union of a narrow Reader and a narrow Writer.
The combined `*RepositoryInterface` still exists (extends both), so
drivers' implementations don't have to change and `Wiring`'s typehints
stay maximally permissive at the composition boundary.

The value lands when consumers — both internal collaborators and
driver-side reactions — typehint the narrowest role they actually need.
Webhook reactions that only persist state can typehint
`SubscriptionWriter` instead of pulling in the read surface they never
use. Read-only consumers (dashboards, status checks) can typehint
`SubscriptionReader`. Mocks get smaller.

New contracts (all in Vatly\Fluent\Contracts):

- CustomerReader     / CustomerWriter
- SubscriptionReader / SubscriptionWriter
- OrderReader        / OrderWriter

The existing three combined interfaces are now thin extensions:
`CustomerRepositoryInterface extends CustomerReader, CustomerWriter`
etc.

Internal narrowings:

- SubscriptionHandle::\$subscriptions → SubscriptionWriter
  (only calls update())
- Billable::\$orders                  → OrderReader
  (Billable never writes orders; reactions own that path)
- Billable::\$customers               → CustomerWriter
  (only calls save() during customer creation)
- BillableFactory matches Billable's narrower hints

Webhook reactions and WebhookProcessorFactory keep the combined
interface — the built-in reactions genuinely need both read and write
(find existing → store-or-update).

WebhookCallRepositoryInterface is left untouched: it has no read methods,
so a Reader/Writer split would add an empty interface for no benefit.

Tests: 149/149 pass (no test changes needed — mocks of the combined
interface satisfy the narrower role hints via standard variance).
phpstan: clean.

README updated:
- Step 5 of the driver-build guide flags the Reader/Writer split.
- "Contracts at a glance" section names each split.
