+++ src/Billable.php
@@ -18,9 +18,9 @@ use Vatly\Fluent\Builders\CheckoutBuilder;
 use Vatly\Fluent\Builders\SubscriptionBuilder;
 use Vatly\Fluent\Contracts\BillableInterface;
 use Vatly\Fluent\Contracts\ConfigurationInterface;
-use Vatly\Fluent\Contracts\CustomerRepositoryInterface;
+use Vatly\Fluent\Contracts\CustomerWriter;
 use Vatly\Fluent\Contracts\OrderInterface;
-use Vatly\Fluent\Contracts\OrderRepositoryInterface;
+use Vatly\Fluent\Contracts\OrderReader;
 use Vatly\Fluent\Contracts\SubscriptionInterface;
 use Vatly\Fluent\Contracts\SubscriptionRepositoryInterface;
 use Vatly\Fluent\Exceptions\CustomerAlreadyCreatedException;
@@ -37,8 +37,8 @@ class Billable
     public function __construct(
         private BillableInterface $owner,
         private SubscriptionRepositoryInterface $subscriptions,
-        private CustomerRepositoryInterface $customers,
-        private OrderRepositoryInterface $orders,
+        private CustomerWriter $customers,
+        private OrderReader $orders,
         private ConfigurationInterface $config,
         private CreateCheckout $createCheckoutAction,
         private CreateCustomer $createCustomerAction,
